change static text to read from messages.properties (#DEV-85)

+++ src/main/java/sg/ncl/testbed_interface/Experiment.java
@@ -68,7 +68,7 @@ public class Experiment {
         this.experimentOwnerId = experimentOwnerId;
     }
     
-    public long getTeamId() {
+    public int getTeamId() {
         return teamId;
     }
